Add support for RPi4 (#7)
* merge `struct RpiError` and `enum Kind` into one `enum RpiError`

The error type is currently a rudimentary newtype over an enum. Since
there don't seem to be any uses of this (a once-planned rich error
type?), simplify it into one type.

* add support for RPi4

* comment out `overlay` and `port` options

Since these are not yet used, comment them out in the same manner as
options `directory` and `signed` just to not generate warnings.

* Cargo.toml: remove `dd-lib` dependency

Not only this crate is unused in this project, but it also fails to
build on 32-bit platforms because of `usize` multiplication overflows.

* fix `second_stage_boot` sending incorrect initial boot_message

In the first transfer, second_stage_boot() needs to send a struct
containing expected size and boot signature (if used). Without this,
the device will not respond to reads, and the subsequent ep_read will
timeout.

* fix missing condition causing premature loop exit

* fix some clippy warnings and prettify format strings

* change visibility of Options' values and fix not sleeping enough

In the original version, usleep(delay) is performed after any two of the
failure points of Initialize_Device(): after an uneventful call to
open_device_with_vid() (analogous to get_device()), or failure to claim
USB device. Failing to open device also incurs a usleep(200).

In this version, usleep(delay) is only performed after failing to
open/get device. Fix this, and don't forget the original 200 usecs.

* bump `rusb` version